Fix for array fields. Add additional tests for float developer fields.

+++ scripts/gen_profile.py
@@ -2,12 +2,11 @@ import argparse
 import inflection
 import jinja2
 import os
+from fit_tool import SDK_VERSION
+from fit_tool.base_type import BaseType
+from fit_tool.field import Field
+from fit_tool.gen.profile import Profile, Message
 from pathlib import Path
-
-from fit import SDK_VERSION
-from fit.base_type import BASE_TYPE_BY_NAME, BaseType
-from fit.field import Field
-from fit.profile.profile import Profile, Message
 
 DEFAULT_BUILD_PATH = './build'
 
@@ -24,7 +23,6 @@ EXAMPLE:
     gen_profile.py build
 
 """)
-    # parser.add_argument("build_dir")
 
     return parser.parse_args()
 
@@ -46,6 +44,8 @@ RESERVED_WORDS = ['new', 'switch', 'enum']
 
 
 def convert_value_name(name: str) -> str:
+    name = name.replace('wkt_', 'workout_')
+
     if name[0].isdigit():
         name = 'n' + name
     name = inflection.camelize(name, uppercase_first_letter=False)
@@ -64,7 +64,7 @@ def field_name_to_class_name(message: Message, field: Field, name: str) -> str:
     name = name.replace('wkt_', 'workout_')
 
     # if not a common field then prefix the message name so we don't get class naming conflicts
-    if field.def_num not in [253, 254]:
+    if field.field_id not in [253, 254]:
         name = message.name + '_' + name
 
     name = inflection.camelize(name, uppercase_first_letter=True)
@@ -111,62 +111,71 @@ def subfield_name_to_property_name(field: Field, name: str) -> str:
     return name
 
 
-def get_type_name_from_field(field: Field, base_type: BaseType):
-    if base_type.name == 'string':
+def get_type_name_from_field(field: Field):
+    if field.base_type.name == 'STRING':
         return 'String'
-    elif 'float' in base_type.name or (field.scale and field.scale != 1):
+    elif 'float' in field.base_type.name or (field.scale and field.scale != 1):
         return 'double'
     else:
         return 'int'
 
 
 def get_field_property_type_name(profile: Profile, field: Field) -> str:
-    if field.type_:
-        if 'bool' in field.type_.name:
-            return 'bool'
-        elif 'date_time' == field.type_.name:
-            return 'int'
-        elif 'fit_base_type' in field.type_.name or field.type_.base_type.name != 'enum':
-            return get_type_name_from_field(field, field.type_.base_type)
+
+    if field.type_name:
+        if 'bool' in field.type_name:
+            type_name = 'bool'
+        elif 'date_time' == field.type_name:
+            type_name = 'int'
+        elif 'fit_base_type' in field.type_name:
+            type_name = 'int'
+
+        elif field.base_type == BaseType.ENUM:
+            type_name = profile.type_class_name_by_name[field.type_name]
         else:
-            return profile.type_class_name_by_name[field.type_.name]
+            type_name = get_type_name_from_field(field)
     else:
-        return get_type_name_from_field(field, field.base_type)
+        type_name = get_type_name_from_field(field)
+
+    # handle arrays
+    if field.array_type is None:
+        return type_name
+    else:
+        if field.base_type.name == 'BYTE':
+            return 'Uint8List'
+        if type_name == 'String':
+            return 'String'
+        else:
+            return f'List<{type_name}>'
 
 
 def main():
-    # args = parse_args()
-
     build_path = DEFAULT_BUILD_PATH
     profile_path = os.path.join(build_path, 'profile')
     messages_path = os.path.join(profile_path, 'messages')
 
-    messages_dir = Path(messages_path);
-    for p in messages_dir.glob('*.dart'):
-        p.unlink()
-    messages_dir.rmdir()
+    messages_dir = Path(messages_path)
+    if messages_dir.exists():
+        for p in messages_dir.glob('*.dart'):
+            p.unlink()
+        messages_dir.rmdir()
 
-    profile_dir = Path(profile_path);
-    for p in profile_dir.glob('*.dart'):
-        p.unlink()
-    profile_dir.rmdir()
+    profile_dir = Path(profile_path)
+    if profile_dir.exists():
+        for p in profile_dir.glob('*.dart'):
+            p.unlink()
+        profile_dir.rmdir()
 
     Path(profile_path).mkdir(parents=True, exist_ok=True)
     Path(messages_path).mkdir(parents=True, exist_ok=True)
-
-    #
-    # Gen profile_type.dart
-    #
-    env = jinja2.Environment(loader=jinja2.FileSystemLoader('.'), )
-
-    template = env.get_template('templates/profile_type.jinja')
 
     profile = Profile.get_default_profile()
 
     profile.types_by_name.pop('bool')
     profile.types_by_name.pop('fit_base_type')
 
-    profile_types = {un_title(convert_type_name(k)) for k in profile.types_by_name.keys()}
+    profile_types = list({un_title(convert_type_name(k)) for k in profile.types_by_name.keys()})
+    profile_types.sort()
 
     profile.type_class_name_by_name = {k: convert_type_name(k) for k in
                                        profile.types_by_name.keys()}
@@ -183,7 +192,7 @@ def main():
         for field in message.fields_by_name.values():
             field.subfield_property_name_by_name = {}
             field.subfield_property_type_by_name = {}
-            for subfield in field.subfields:
+            for subfield in field.sub_fields:
                 ref_field_key = list(subfield.ref_field_map.keys())[0]
                 subfield.ref_field = message.fields_by_name[ref_field_key]
                 subfield.ref_field_values = subfield.ref_field_map[ref_field_key]
@@ -198,7 +207,13 @@ def main():
         profile_type.values_by_name = {convert_value_name(k): v for k, v in
                                        profile_type.values_by_name.items()}
 
-    rendering = template.render(base_type=BASE_TYPE_BY_NAME, profile=profile,
+    env = jinja2.Environment(loader=jinja2.FileSystemLoader('.'), )
+
+    #
+    # Gen profile_type.py
+    #
+    template = env.get_template('templates/profile_type.jinja')
+    rendering = template.render(profile=profile,
                                 profile_types=profile_types,
                                 sdk_version=SDK_VERSION)
     filename = os.path.join(profile_path, 'profile_type.dart')
@@ -207,10 +222,11 @@ def main():
 
     for name, message in profile.messages_by_name.items():
         template = env.get_template('templates/message.jinja')
-        rendering = template.render(base_type=BASE_TYPE_BY_NAME, profile=profile,
+        rendering = template.render(profile=profile,
                                     sdk_version=SDK_VERSION,
                                     class_name=inflection.camelize(message.name + '_message'),
-                                    message=message)
+                                    message=message,
+                                    BaseType=BaseType)
         filename = os.path.join(messages_path,
                                 f"{inflection.underscore(message.name + '_message')}.dart")
         with (open(filename, 'w')) as file_out:
